♻️ refactor state management for FilterChips
Why? Since the form is a bit more "complex" in hiding fields based on
other fields, it felt apropriate to "somewhat centralize" the state,
not going full context here, but sending down the useState() is fine
for now in terms of complexity imo.

Unfortunately the smarter / optimized behaviour has been disabled for
now, so no prefetching of the resulting query when hovering over a
FilterChip. It also requires you to hit enter or click the search icon
in order to trigger a search. It's more manual, but simpler to
understand (and to develop). Consider re-adding smarts. The smarts was
also not "fully smart" in its previous incarnation, since we couldn't
have the search be "instant and automatic" (not requiring any
enter/clicks beyond typing).