semver search in changelog entries (#3989)
* :memo: WIP setup for semver search in changelog entries

* settings display & opt-out

* maxWidth to not break mobile

* make feature opt-in

* now support semver range searches

* join filter back to valid semver range syntax

* tweaks to search syntaks Popover and button

* use Helptext instead of custom Helptext

* checkbox for search preference

* default checked for snappy prefilled value from search params

* group all form into SearchForm for use client shennanigans and state management

* typing++

* ♻️ refactor state management for FilterChips

Why? Since the form is a bit more "complex" in hiding fields based on
other fields, it felt apropriate to "somewhat centralize" the state,
not going full context here, but sending down the useState() is fine
for now in terms of complexity imo.

Unfortunately the smarter / optimized behaviour has been disabled for
now, so no prefetching of the resulting query when hovering over a
FilterChip. It also requires you to hit enter or click the search icon
in order to trigger a search. It's more manual, but simpler to
understand (and to develop). Consider re-adding smarts. The smarts was
also not "fully smart" in its previous incarnation, since we couldn't
have the search be "instant and automatic" (not requiring any
enter/clicks beyond typing).

* typing

---------

Co-authored-by: Julian Nymark <julian.nymark@nav.no>